requirements.txt
# HTTP requests (Alpha Vantage API)
requests==2.31.0

# RabbitMQ client
pika==1.3.2

# Data processing & analytics
pandas==2.1.4
numpy==1.26.2

mysql-connector-python==8.3.0

# Date & time handling
python-dateutil==2.8.2

# Environment variable management
python-dotenv==1.0.0

# Logging
loguru==0.7.2
